separate option parsing from generator class construction
